Move default context to context processor
