+++ iquise/utils.py
@@ -33,26 +33,6 @@ def decode_data(hash_, enc):
     data = pickle.loads(zlib.decompress(text.decode('base64')))
     return data
 
-def basic_context(request):
-    staff_reg_url = None
-    if request.user.is_superuser:
-        expires = (timezone.localtime()+timedelta(days=1)).strftime('%x%X')
-        staff_reg_url = reverse('members:register',args=[''.join(encode_data(expires))])
-    notifications = []
-    if settings.DEBUG:
-        notifications = ['DEVELOPMENT SERVER']
-    extra_notification = request.session.pop("extra_notification", None)
-    if extra_notification:
-        notifications.append(extra_notification)
-    # No analytics if superuser
-    if request.user.is_superuser:
-        useAnalytics = False
-    else:
-        useAnalytics = not settings.DEBUG
-    iquise = IQUISE.objects.first() # Returns none if doesn't exist
-    donors = [unicode(d) for d in Donor.objects.all()]
-    return {'iquise':iquise,'useAnalytics': useAnalytics,'notifications':notifications,'donors':donors,'staff_reg_url':staff_reg_url}
-
 class AlwaysClean(models.Model):
     def save(self,*args,**kwargs):
         self.full_clean()
