+++ website/templates/home/base.html
@@ -78,6 +78,13 @@
                                 <div><h2>{{notification}}</h2></div>
                             </div>
                             {% endfor %}
+                            {% for notification in more_notifications %}
+                            <!-- Notification -->
+                            <div class="notification">
+                                <a href="#" class='close'>&times;</a>
+                                <div><h2>{{notification}}</h2></div>
+                            </div>
+                            {% endfor %}
 
                             <!-- Banner -->
                             {% block banner %}{% endblock %}
